requirements.txt
pandas
pyarrow
duckdb
requests
python-dotenv
tqdm
PyYAML